Refactor: components 하위 폴더 정리
