+++ src/components/Home/List/DiaryItem.module.scss
@@ -1,5 +1,5 @@
-@import '../../styles/variables/colors.scss';
-@import '../../styles/variables/fonts.scss';
+@import '../../../styles/variables/colors.scss';
+@import '../../../styles/variables/fonts.scss';
 
 .DiaryItem {
   padding: 16px 16px 16px 16px;
